refactor(config): add gson instance in GameConfig, as well as moving it to its own package
